feat: Add new class and method to decoration and output message. Add comments to quick doc.
