+++ index-files/index-1.html
@@ -1,0 +1,67 @@
+<!DOCTYPE HTML>
+<html lang="ru">
+<head>
+<!-- Generated by javadoc (17) on Sat Mar 26 21:00:19 MSK 2022 -->
+<title>C-Index</title>
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+<meta name="dc.created" content="2022-03-26">
+<meta name="description" content="index: C">
+<meta name="generator" content="javadoc/IndexWriter">
+<link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
+<link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
+<link rel="stylesheet" type="text/css" href="../jquery-ui.overrides.css" title="Style">
+<script type="text/javascript" src="../script.js"></script>
+<script type="text/javascript" src="../script-dir/jquery-3.5.1.min.js"></script>
+<script type="text/javascript" src="../script-dir/jquery-ui.min.js"></script>
+</head>
+<body class="index-page">
+<script type="text/javascript">var pathtoroot = "../";
+loadScripts(document, 'script');</script>
+<noscript>
+<div>JavaScript is disabled on your browser.</div>
+</noscript>
+<div class="flex-box">
+<header role="banner" class="flex-header">
+<nav role="navigation">
+<!-- ========= START OF TOP NAVBAR ======= -->
+<div class="top-nav" id="navbar-top">
+<div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
+<ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
+<li><a href="../package-summary.html">Package</a></li>
+<li>Class</li>
+<li><a href="../overview-tree.html">Tree</a></li>
+<li class="nav-bar-cell1-rev">Index</li>
+<li><a href="../help-doc.html#index">Help</a></li>
+</ul>
+</div>
+<div class="sub-nav">
+<div class="nav-list-search"><label for="search-input">SEARCH:</label>
+<input type="text" id="search-input" value="search" disabled="disabled">
+<input type="reset" id="reset-button" value="reset" disabled="disabled">
+</div>
+</div>
+<!-- ========= END OF TOP NAVBAR ========= -->
+<span class="skip-nav" id="skip-navbar-top"></span></nav>
+</header>
+<div class="flex-content">
+<main role="main">
+<div class="header">
+<h1>Index</h1>
+</div>
+<a href="index-1.html">C</a>&nbsp;<a href="index-2.html">P</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
+<h2 class="title" id="I:C">C</h2>
+<dl class="index">
+<dt><a href="../com.tcs.edu.printer.ConsolePrinter.html" class="type-name-link" title="class in Unnamed Package">com.tcs.edu.printer.ConsolePrinter</a> - Class in <a href="../package-summary.html">Unnamed Package</a></dt>
+<dd>
+<div class="block">Класс com.tcs.edu.printer.ConsolePrinter
+ Включает в себя методы связанные с выводом сообщений в консоль</div>
+</dd>
+<dt><a href="../com.tcs.edu.printer.ConsolePrinter.html#%3Cinit%3E()" class="member-name-link">com.tcs.edu.printer.ConsolePrinter()</a> - Constructor for class <a href="../com.tcs.edu.printer.ConsolePrinter.html" title="class in Unnamed Package">com.tcs.edu.printer.ConsolePrinter</a></dt>
+<dd>&nbsp;</dd>
+</dl>
+<a href="index-1.html">C</a>&nbsp;<a href="index-2.html">P</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a></main>
+</div>
+</div>
+</body>
+</html>
