+++ src/main/java/com/tcs/edu/printer/ConsolePrinter.java
@@ -1,0 +1,17 @@
+package com.tcs.edu.printer;
+
+/** Класс com.tcs.edu.printer.ConsolePrinter
+ * Включает в себя методы связанные с выводом сообщений в консоль
+ *
+ */
+public class ConsolePrinter {
+
+    /** Метод print
+     * Предназначен для вывода сообщений в консоль
+     * @param message текстовый параметр с сообщением для вывода в консоль
+     * @author m.petrukhin
+     */
+    public static void print(String message) {
+        System.out.println(message);
+    }
+}
